Rewrite the makefile and support non-static targets
This commit rewrites the Makefile in an effort to make it easier to add
complexity in the future.

The primary change is that the bulk of the compiling now made via the
implicit rules.

It is expected for this commit to generate some problems with older
versions of `gcc` or `msvc`.

To build a non-static ldpl interpreter use something like:
	make clean all STATIC=n

Any other value to the STATIC variable builds a static executable.