Advise Windows user to change git core.autoclrf config
Windows users should be wary that, DOS-style end-of-line (CLRF) does
not produce the expected output when running the awk script
(lib-to-string.awk).
This causes compilation errors further down.

One way of remediation, which is showcased here, is to set git
`core.autoclrf` setting to `input`. This will make sure that files will
be checked out with Unix-style end-of-line (LF), preventing further
annoyances.

Alternatively, users can use a program such as `dos2unix` on directly on
source files before building the project.